Updated the omssa-parser library to version 1.3.7. Removed a System.out.println in SequestSpectrumMillConverter.

+++ pride_converter/prideconverter-core/src/main/java/no/uib/prideconverter/OmssaConverter.java
@@ -74,7 +74,7 @@ public class OmssaConverter {
         ArrayList<CvParam> cVParams;
         ArrayList<UserParam> userParams;
         List<Integer> mzValues, intensityValues;
-        int omssaAbundanceScale;
+        long omssaAbundanceScale;
         Collection<SpectrumDescComment> spectrumDescriptionComments;
 
         MSModHit currentMSModHit;
